Need check string

+++ lab2/main.c
@@ -5,7 +5,7 @@ int main()
 	mt_setbgcolor(clr_blue);
 	printf("LOL\n");
     mt_setbgcolor(clr_default);
-    uint16_t y, x;
+    int y, x;
     mt_getscreensize(&y, &x);
     mt_gotoXY((uint16_t) (x / 2), (uint16_t) (y / 2));
     return 0;
